requirements.txt
aiogram==3.10.0
aiohttp==3.9.5   # совместимо с aiogram 3.10.0
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
